+++ FOSSDB_web/apps/fossdb/views.py
@@ -1,8 +1,8 @@
 from django.contrib.auth.decorators import login_required, permission_required
 from django.shortcuts import redirect, render
+from programming_language.forms import ProjectProgrammingLanguageFormSet
 
-from .forms import (ProjectForm, ProjectHostingPlatformFormSet,
-                    ProjectProgrammingLanguageFormSet)
+from .forms import ProjectForm, ProjectHostingPlatformFormSet
 from .models import Project
