+++ FOSSDB_web/apps/programming_language/admin.py
@@ -1,0 +1,11 @@
+from django.contrib import admin
+
+from .models import ProgrammingLanguage, ProjectProgrammingLanguage
+
+
+class ProjectProgrammingLanguageInline(admin.TabularInline):
+    model = ProjectProgrammingLanguage
+    extra = 1
+
+
+admin.site.register(ProgrammingLanguage)
